% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-halfspacemass-def.R
\name{add_halfspaces}
\alias{add_halfspaces}
\title{Add Halfspaces}
\usage{
add_halfspaces(complete_data, n_halfspace, input)
}
\arguments{
\item{complete_data}{A data set with all containing required unique
observations to \code{\link[=train_depth]{train_depth()}}}

\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{input}{the parameter set an object was trained on}
}
\value{
a halfspaces object of n halfspace objects
}
\description{
This function calculates \code{n_halfspaces} new halfspaces from
\code{complete_data}
}
\seealso{
\code{\link[=update.halfspaces]{update.halfspaces()}}, \code{\link[=train_depth]{train_depth()}}
}
