% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-halfspacemass-def.R
\name{autoplot.halfspaces}
\alias{autoplot.halfspaces}
\title{Plot 2D-Halfspace depth/mass}
\usage{
\method{autoplot}{halfspaces}(
  object,
  data,
  metric = c("mass", "depth"),
  points = TRUE,
  type = c("heatmap", "contour"),
  grid = NULL,
  gridlength = 70L
)
}
\arguments{
\item{object}{a halfspaces object containing a list arbitrary length of
halfspace objects.}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{metric}{default = "mass", alternative "depth"}

\item{points}{logical; should datapoints be plotted? default = TRUE}

\item{type}{char; choice between "heatmap" and "contour lines"}

\item{grid}{numeric dataframe or matrix defining the grid to plot at;
default = NULL}

\item{gridlength}{integer indicating grid length, to compute a grid on,
is ignored if \code{grid} is not NULL; default =  70L}
}
\description{
Autoplot-Method to \code{\link[=train_depth]{train_depth()}} for halfspaces objects.
Plot Halfspace depth/mass in two-dimensional grid\preformatted{Method generates a additive ggplot object by 2 Steps:
* 1st Step [design_grid()]
Creates grid, predicts on it and saves it as database in a ggplot object.
* 2nd Step [add_tile()] or [add_contour()] and/or [add_points()]
Adds layers objects tile, contour and points to the grid, by argument
specification
* 3rd Step Labs 'n beauty
Add lab aspects depending on `metric` for intelligibility and theme for
beauty to the grid
}
}
\section{Functions}{
\itemize{
\item \code{autoplot.halfspaces}: ggplot2 object; either a heatmap od a
contour-line-plot in range of the grid
}}

\examples{
    library(halflinger)

    data <- matrix(c(rnorm(100), rnorm(100, 1, 5)), ncol = 2)
    train_data <- train_depth(data, n_halfspace = 100, seed = 123)
    \dontrun{autoplot.halfspaces(train_data, data, type = "contour")}

}
\seealso{
\preformatted{[train_depth()], [update.halfspaces()], [predict.halfspaces()]
}
}
