% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{compute_mass}
\alias{compute_mass}
\title{Mass above split}
\usage{
compute_mass(projections, split)
}
\arguments{
\item{projections}{vector: projections of <data> on <direction> in units of
"length of <direction>"}

\item{split}{"offset" of the halfspace boundary along its normal vector}
}
\value{
a vector with relative data frequency "at or above" the boundary
defined by split
}
\description{
get proportion of observations above boundary separating 2
halfspaces
}
