% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-halfspacemass-def.R
\name{design_grid}
\alias{design_grid}
\alias{create_grid}
\title{Design the grid}
\usage{
design_grid(grid, data, gridlength = 70L)

create_grid(data, gridlength = 70L)
}
\arguments{
\item{grid}{numeric dataframe or matrix defining the grid to plot at;
default = NULL}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{gridlength}{integer indicating grid length, to compute a grid on,
is ignored if \code{grid} is not NULL; default =  70L}
}
\description{
Designs a 2 column grid as database for a gg_plot object, if
\code{grid} is not supplied, generates a grid by \code{gridlength}
}
\section{Functions}{
\itemize{
\item \code{design_grid}: assigns a two dimensional grid as database of a
ggplot object

\item \code{create_grid}: creates a grid of \code{gridlength}x\code{gridlength} Pixels,
where the range of the each axis corresponds to the range of the \code{data}
columns. The first column is plotted on the x-axis, the second on the
y-axis.
}}

