% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-testfunctions.R
\name{generate_train_data}
\alias{generate_train_data}
\alias{generate_plot_faith}
\title{Generate test data}
\usage{
generate_train_data(
  n_halfspace = 1000,
  subsample = 1,
  scope = 1,
  seed = NULL,
  scale = FALSE,
  seed0 = 121133
)

generate_plot_faith(
  n_halfspace = 10000,
  subsample = 1,
  scope = 1,
  seed = NULL,
  scale = FALSE,
  metric = c("mass", "depth"),
  points = TRUE,
  type = c("heatmap", "contour"),
  grid = NULL,
  gridlength = 70L
)
}
\arguments{
\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{subsample}{what proportion of data to use for each halfspace
estimation, default = 1}

\item{scope}{minimum 1; controls size of region of convexity for halfspace
mass (\eqn{\lambda} in the paper) i.e., how far outside of sampled data range
the sampled hyperplanes can lie; default = 1}

\item{seed}{optional RNG seed; default = NULL}

\item{scale}{logical; wether data should be scaled; default = FALSE}

\item{seed0}{an additional RNG seed}

\item{metric}{default = "mass", alternative "depth"}

\item{points}{logical; should datapoints be plotted? default = TRUE}

\item{type}{char; choice between "heatmap" and "contour lines"}

\item{grid}{numeric dataframe or matrix defining the grid to plot at;
default = NULL}

\item{gridlength}{integer indicating grid length, to compute a grid on,
is ignored if \code{grid} is not NULL; default =  70L}
}
\description{
This functions are helpers to unit tests.
}
\section{Functions}{
\itemize{
\item \code{generate_train_data}: generates datasets in order to approve, the
functionality of \code{\link[=train_depth]{train_depth()}}, \code{\link[=update.halfspaces]{update.halfspaces()}},
\code{\link[=predict.halfspaces]{predict.halfspaces()}}

\item \code{generate_plot_faith}: possibility to tune over all defaults of
tain_depth and autoplot.halfspaces to approve the functionality of
\code{\link[=autoplot.halfspaces]{autoplot.halfspaces()}}
}}

