% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{get_directions}
\alias{get_directions}
\title{Draw directions in <dim>-dimnesional Space}
\usage{
get_directions(n_halfspace, dims = 2)
}
\arguments{
\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{dims}{dimensions of space to draw directions from}
}
\value{
a <dims> x <n_halfspace> matrix of <n_halfspace> <dims>-dimensional
directions
}
\description{
Uniformly sample <n_halfspace> <dim>-dimensional direction
vectors
}
