% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{get_halfspace}
\alias{get_halfspace}
\title{Create a Halfspace}
\usage{
get_halfspace(normal, data, subsample, scope)
}
\arguments{
\item{normal}{a vector of indicating a direction in space}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{subsample}{what proportion of data to use for each halfspace
estimation, default = 1}

\item{scope}{minimum 1; controls size of region of convexity for halfspace
mass (\eqn{\lambda} in the paper) i.e., how far outside of sampled data
range the sampled hyperplanes can lie; default = 1}
}
\value{
a halfspace object containing a direction, a splitpoint and a
measure of mass
}
\description{
Project a subsample of \code{data} on
the direction, create a splitpoint in the projection and measure the mass
above the split point vectors
}
