% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-halfspacemass-def.R
\name{get_mass}
\alias{get_mass}
\alias{get_depth}
\title{Calculate Halfspaces}
\usage{
get_mass(data, object)

get_depth(data, object)
}
\arguments{
\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{object}{a \code{\link[=halfspaces]{halfspaces()}} object containing a list arbitrary length of
\code{\link[=halfspace]{halfspace()}} objects.}
}
\description{
These functions calculate a prediction on depth/mass of data
}
\section{Functions}{
\itemize{
\item \code{get_mass}: numeric vector with halfspace mass values for data
for each combination of a halfspace and a data point, the algorithm
uses the mass on the side of the split which the data point lies
on, standardizes the input and takes the mean

\item \code{get_depth}: numeric vector with halfspace depth values for data
initialized in 1 the algorithm iteratively searches in each halfspace
for the minimal masses of each observation
}}

\seealso{
\code{\link[=predict.halfspaces]{predict.halfspaces()}}
}
