% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-S3-halfspace.R
\name{halfspace}
\alias{halfspace}
\alias{new_halfspace}
\alias{is_halfspace}
\alias{validate_halfspace}
\title{Define S3-class halfspace}
\usage{
halfspace(normal, split, mass_above, col_data)

new_halfspace(normal, split, mass_above)

is_halfspace(object, col_data)

validate_halfspace(object, col_data)
}
\arguments{
\item{normal}{a vector indicating a direction in space}

\item{split}{output from \code{\link[=sample_split]{sample_split()}}}

\item{mass_above}{a vector with relative data frequency "at or above" the
split}

\item{col_data}{dimesions of the space that halfspace lives in}

\item{object}{a object of class halfspace}
}
\description{
Define a new S3 class \code{halfspace}.
Structure of class definition:
\itemize{
\item interface
\item constructor
\item validator
}
}
\section{Functions}{
\itemize{
\item \code{halfspace}: a halfspace object containing a direction, a splitpoint
and a measure of mass \code{\link[=train_depth]{train_depth()}}, \code{\link[=update.halfspaces]{update.halfspaces()}},
\code{\link[=predict.halfspaces]{predict.halfspaces()}}

\item \code{new_halfspace}: a low level constructor for class halfspace

\item \code{is_halfspace}: check on structur of halfspace object w.r.t data

\item \code{validate_halfspace}: validate structure of halfspace object
}}

\seealso{
\code{\link[=get_halfspace]{get_halfspace()}}, \code{\link[=halfspaces]{halfspaces()}}
}
