% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-S3-halfspaces.R
\name{halfspaces}
\alias{halfspaces}
\alias{new_halfspaces}
\alias{validate_halfspaces}
\title{Define S3-class halfspaces}
\usage{
halfspaces(object, train_data, subsample, scope, seed, scale)

new_halfspaces(object, train_data, subsample, scope, seed, scale)

validate_halfspaces(object)
}
\arguments{
\item{object}{list of arbitrary length containing \code{halfspace} objects}

\item{train_data}{input data in \code{\link[=train_depth]{train_depth()}}}

\item{subsample}{what proportion of data to use for each \code{\link[=halfspace]{halfspace()}} estimation, default = 1}

\item{scope}{minimum 1; controls size of region of convexity for \code{\link[=halfspace]{halfspace()}}
mass (\eqn{\lambda} in the paper) i.e., how far outside of sampled data
range the sampled hyperplanes can lie; default = 1}

\item{seed}{optional RNG seed; default = NULL}

\item{scale}{logical; wether data should be scaled; default = FALSE}
}
\description{
Define a new S3 class \code{halfspaces}.
Structure of class definition:
\itemize{
\item interface
\item constructor
\item validator
}
}
\section{Functions}{
\itemize{
\item \code{halfspaces}: Attributes of this objects are all neccesary arguments from
\code{\link[=train_depth]{train_depth()}} to reproduce the sample if a seed is supplied

\item \code{new_halfspaces}: a low level constructor for class halfspaces

\item \code{validate_halfspaces}: validate structure of halfspaces object
}}

\seealso{
\code{\link[=train_depth]{train_depth()}}, \code{\link[=halfspace]{halfspace()}}, \code{\link[=update.halfspaces]{update.halfspaces()}},
\code{\link[=autoplot.halfspaces]{autoplot.halfspaces()}}, \code{\link[=predict.halfspaces]{predict.halfspaces()}}
}
