% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-halfspacemass-def.R
\name{predict.halfspaces}
\alias{predict.halfspaces}
\title{Test algorithm of Chen et al.}
\usage{
\method{predict}{halfspaces}(object, data, metric = c("mass", "depth"), ...)
}
\arguments{
\item{object}{a \code{\link[=halfspaces]{halfspaces()}} object containing a list arbitrary length of
\code{\link[=halfspace]{halfspace()}} objects.}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{metric}{default = "mass", alternative "depth"}

\item{...}{ellipsis}
}
\value{
numeric vector with depth metric values for data: either halfspace
mass or halfspace depth
}
\description{
Predict-Method to \code{\link[=train_depth]{train_depth()}} for \code{\link[=halfspaces]{halfspaces()}} objects.
Evaluates halfspace-mass or (approximate) -depth for points in \code{data}

Functionality splits by \code{metric} in in two main scenarios:
\itemize{
\item Scenario \code{\link[=get_mass]{get_mass()}}:
\code{metric = "mass"} computes approximate halfspace mass
\item Scenario \code{\link[=get_depth]{get_depth()}}:
\code{metric = "depth"} computes approximate halfspace depth ("Tukey depth")
}
}
\examples{
library(halflinger)

data <- matrix(c(rnorm(100), rnorm(100, 1, 5)), ncol = 2)
train_data <- train_depth(data[51:100, ], n_halfspace = 100, seed = 123)
predict(train_data, data[1:50, ])
}
\references{
Chen, B., Ting, K.M., Washio, T. et al.,
Mach Learn (2015): 100(2):677--699
Half-space mass a maximally robust and efficient data depth method
\url{https://doi.org/10.1007/s10994-015-5524-x}
}
\seealso{
\code{\link[=train_depth]{train_depth()}}, \code{\link[=update.halfspaces]{update.halfspaces()}}, \code{\link[=autoplot.halfspaces]{autoplot.halfspaces()}}
}
