% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to check and prepare \code{data} for calculation}
\usage{
prepare_data(data, scale = FALSE)
}
\arguments{
\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{scale}{logical; wether data should be scaled; default = FALSE}
}
\value{
a numeric (scaled) matrix
}
\description{
ensures \code{data} to be a numeric matrix & scales them if \code{scale}
}
\seealso{
\code{\link[=train_depth]{train_depth()}}, \code{\link[=update.halfspaces]{update.halfspaces()}}, \code{\link[=autoplot.halfspaces]{autoplot.halfspaces()}} and
\code{\link[=predict.halfspaces]{predict.halfspaces()}}
}
