% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{project_scalar}
\alias{project_scalar}
\title{Scalar projection}
\usage{
project_scalar(data, normal)
}
\arguments{
\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{normal}{a vector of indicating a direction in space}
}
\value{
vector: projections of <data> on <direction> in units of
"length of <direction>"
}
\description{
project data points onto the normal vector defining the orientation
of a halfspace
}
\seealso{
\url{https://en.wikipedia.org/wiki/Scalar_projection}
}
