% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{sample_split}
\alias{sample_split}
\title{Draw a Splitpoint}
\usage{
sample_split(projections, scope)
}
\arguments{
\item{projections}{vector: projections of \code{data} on \code{normal} in units of
"length of \code{normal}"}

\item{scope}{minimum 1; controls size of region of convexity for halfspace
mass (\eqn{\lambda} in the paper) i.e., how far outside of sampled data
range the sampled hyperplanes can lie; default = 1}
}
\value{
"offset" of the halfspace boundary along its normal vector
}
\description{
sample a splitpoint in span-range of data
}
