% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{subsample}
\alias{subsample}
\title{Create a Subsample}
\usage{
subsample(data, subsample = 1)
}
\arguments{
\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{subsample}{what proportion of data to use for each halfspace
estimation, default = 1}
}
\value{
a subsample of rows in data
}
\description{
creates a subsample of <data>, containing a <subsample>
proportion of rows in data
}
