% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-halfspacemass-def.R
\name{add_data}
\alias{add_data}
\title{Add data}
\usage{
add_data(train_data, data, add)
}
\arguments{
\item{train_data}{the dataset a concerning \code{object} was trained on}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{add}{logical, indicating wether \code{data} should be added to \code{train_data}
from \code{object}}
}
\value{
A data set with all containing required unique observations to
\code{\link[=train_depth]{train_depth()}}
}
\description{
This function takes the observations from train data and the
observations from input data and returns their union amount if add is set to
TRUE; in addtion add_data formats the \code{data}.
}
