% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halflinger-package.R
\docType{package}
\name{halflinger-package}
\alias{halflinger}
\alias{halflinger-package}
\title{halflinger: Package to train, predict, update  and plot Halfspacemass and Turkeydetph}
\description{
This package implements "Halfspace Mass" (a data depth) as well as a
    Monte Carlo-approximate halfspace (Tukey) depth. Therefore it provides a 
    train function that creates an arbitrary amount of Halfspaces from given 
    data and extends update, predict and autoplot generics by halfspace methods.
}
\details{
\preformatted{For further information brows main functions
[train_depth()], [halflinger::update.halfspaces()],
[halflinger::predict.halfspaces()], [halflinger::autoplot.halfspaces()].
Otherwise refer to the vignettes to see how to format the documentation.
}
}
\references{
\preformatted{Chen, B., Ting, K.M., Washio, T. et al.,
Mach Learn (2015): 100(2):677--699
Half-space mass a maximally robust and efficient data depth method
\url{https://doi.org/10.1007/s10994-015-5524-x}
}
}
\author{
\strong{Maintainer}: Henri Funk \email{henrifnk@gmail.com}

Authors:
\itemize{
  \item Fabian Scheipl \email{fabian.scheipl@stat.uni-muenchen.de}
}

}
\keyword{internal}
