% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-S3-halfspace.R
\name{halfspace}
\alias{halfspace}
\alias{new_halfspace}
\alias{is_halfspace}
\alias{validate_halfspace}
\title{Define Class halfspace}
\usage{
halfspace(normal, split, mass_above, col_data)

new_halfspace(normal, split, mass_above)

is_halfspace(object, col_data)

validate_halfspace(object, col_data)
}
\arguments{
\item{normal}{a vector of indicating a direction in space}

\item{split}{"offset" of the halfspace boundary along its normal vector}

\item{mass_above}{a vector with relative data frequency "at or above" the
split}

\item{col_data}{dimesions of the space that halfspace lives in}

\item{object}{a object of class halfspace}
}
\description{
Define a new S3 class \code{halfspace}.
Structure of class definition: interface, constructor, validator.
}
\section{Functions}{
\itemize{
\item \code{halfspace}: a halfspace object containing a direction, a splitpoint
and a measure of mass \code{\link[=train_depth]{train_depth()}}, \code{\link[=update.halfspaces]{update.halfspaces()}},
\code{\link[=predict.halfspaces]{predict.halfspaces()}}

\item \code{new_halfspace}: a low level constructor for class halfspace

\item \code{is_halfspace}: check on structur of halfspace object w.r.t data

\item \code{validate_halfspace}: validate structur of halfspace object
}}

\seealso{
\code{\link[=get_halfspace]{get_halfspace()}}, \code{\link[=halfspaces]{halfspaces()}}
}
