% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-halfspacemass-def.R
\name{train_depth}
\alias{train_depth}
\title{Implements Training algorithm of Chen et al.}
\usage{
train_depth(
  data,
  n_halfspace = 1000,
  subsample = 1,
  scope = 1,
  seed = NULL,
  scale = FALSE
)
}
\arguments{
\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{subsample}{what proportion of data to use for each halfspace
estimation, default = 1}

\item{scope}{minimum 1; controls size of region of convexity for halfspace
mass (\eqn{\lambda} in the paper) i.e., how far outside of sampled data
range the sampled hyperplanes can lie; default = 1}

\item{seed}{optional RNG seed; default = NULL}

\item{scale}{logical; wether data should be scaled; default = FALSE}
}
\value{
a halfspace object of <n_halfspace> halfspaces, defined by their
normal vector and offset from origin, with estimated data frequencies
above/below halfspace boundary object attributes contain the input
arguments for reproducability
}
\description{
Creates an halfspaces objetct of arbitrary length containing
halfspace objects\preformatted{Functionality after input check can be summed up in two main steps:
* 1st Step [get_directions()]
draw n random directions in m-dimensional space where n is
defined by `n_halfspace` and m is the sum of columns in `data`
* 2nd Step [get_halfspace()]
Iterate over each direction: Project a subsample of `data`
on the direction, create a splitpoint in the projection and measure the
mass above the split point (a direction, a splitpoint and a measure of
mass form a single halfspace)
}
}
\examples{
   library(halflinger)

   data <- matrix(c(rnorm(100), rnorm(100, 1, 5)), ncol = 2)
   train_data <- train_depth(data, n_halfspace = 100, seed = 123)

}
\references{
\preformatted{Chen, B., Ting, K.M., Washio, T. et al.,
Mach Learn (2015): 100(2):677--699
Half-space mass A maximally robust and efficient data depth method
}

\url{https://doi.org/10.1007/s10994-015-5524-x}
}
\seealso{
\preformatted{[update.halfspaces()], [autoplot.halfspaces()], [predict.halfspaces()]
}
}
