% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-halfspacemass-def.R
\name{update.halfspaces}
\alias{update.halfspaces}
\title{update.halfspaces}
\usage{
\method{update}{halfspaces}(
  object,
  data = attr(object, "train_data"),
  n_halfspace = length(object),
  add = FALSE,
  ...
)
}
\arguments{
\item{object}{a halfspaces object containing a list arbitrary length of
halfspace objects.}

\item{data}{numeric matrix or dataframe; \strong{rows} are vectors to project}

\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{add}{logical, indicating wether \code{data} should be added to \code{train_data}
from \code{object}}

\item{...}{ellipsis}
}
\description{
Update-Method to \code{\link[=train_depth]{train_depth()}} for halfspaces objects.
Implementation train algorithm of Chen et al. for adding and changing
\code{data} and amount of \code{halfspaces}.
}
\details{
Functionality after input check can be summed up in 3 main steps:
\subsection{1st Step \code{\link[=add_data]{add_data()}} & \code{\link[=prepare_data]{prepare_data()}}}{

check the input on type and scale & prepare data for calculation
}

\subsection{2nd Step \code{\link[=update_halfspaces]{update_halfspaces()}}}{

update Halfspaces, already calculated in the input \code{object}
}

\subsection{3rd Step \code{\link[=add_halfspaces]{add_halfspaces()}}}{

get new Halfspaces
}
}
\references{
\preformatted{Chen, B., Ting, K.M., Washio, T. et al.,
Mach Learn (2015): 100(2):677--699
Half-space mass a maximally robust and efficient data depth method
\url{https://doi.org/10.1007/s10994-015-5524-x}
}
}
\seealso{
\code{\link[=train_depth]{train_depth()}}, \code{\link[=autoplot.halfspaces]{autoplot.halfspaces()}}, \code{\link[=predict.halfspaces]{predict.halfspaces()}}
}
