% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-halfspacemass-def.R
\name{update_halfspaces}
\alias{update_halfspaces}
\title{Update Halfspaces}
\usage{
update_halfspaces(object, complete_data, n_halfspace, input)
}
\arguments{
\item{object}{a halfspaces object containing a list arbitrary length of
halfspace objects.}

\item{complete_data}{A data set with all containing required unique
observations to \code{\link[=train_depth]{train_depth()}}}

\item{n_halfspace}{number of halfsspaces to draw, default = 1e4}

\item{input}{the parameter set an object was trained on}
}
\value{
A data set that is ready to update \code{\link[=train_depth]{train_depth()}} on
}
\description{
This function updates the already drawn \code{halfspaces} of the
input \code{object}.
}
\details{
Functionality splits in two scenarios:
\subsection{Scenario for an unchanged dataset}{

if dataset remains unchanged, halfspace objects are identical, too
drawn halfspaces are a subset of \code{object}
}

\subsection{Scenario for a changed dataset}{

if dataset changes, new randomly drawn subsamples are needed which provide
a new base for splitpoint criteria and mass_above, therefore whole
<get_halfspaces> needs to be recalled on each direction of the dataset
}
}
